move spaces
